pyproject.toml
[project]
name = "github-explorer"
version = "1.1.0"
description = "A simple MCP GitHub explorer server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.1",
    "requests>=2.32.5",
    "python-dotenv==1.1.0"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling", ]
build-backend = "hatchling.build"

[project.scripts]
server = "server:main"
server_sse = "server_sse:main"
server_shttp = "server_shttp:main"